CV-589 Moved the Request folder classes inside the Models folder. Also changed the namesapce to reflect this.Deleted the Request folder.
