+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web.UnitTests/Controllers/ApprenticesControllerTests/WhenGettingSelectEmployer.cs
@@ -5,7 +5,6 @@ using SFA.DAS.CommitmentsV2.Shared.Interfaces;
 using SFA.DAS.ProviderCommitments.Web.Controllers;
 using SFA.DAS.ProviderCommitments.Web.Models.Apprentice;
 using SFA.DAS.ProviderCommitments.Web.Models.Shared;
-using SFA.DAS.ProviderCommitments.Web.Requests.Apprentice;
 using System.Collections.Generic;
 using System.Threading.Tasks;
