+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web.UnitTests/Validators/NonReservationsAddDraftApprenticeshipRequestValidatorTests.cs
@@ -2,7 +2,7 @@
 using System.Linq.Expressions;
 using FluentValidation.TestHelper;
 using NUnit.Framework;
-using SFA.DAS.ProviderCommitments.Web.Requests;
+using SFA.DAS.ProviderCommitments.Web.Models;
 using SFA.DAS.ProviderCommitments.Web.Validators;
 
 namespace SFA.DAS.ProviderCommitments.Web.UnitTests.Validators
