+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/Cohort/ConfirmEmployerRequest.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.Authorization.ModelBinding;
 
-namespace SFA.DAS.ProviderCommitments.Web.Requests.Cohort
+namespace SFA.DAS.ProviderCommitments.Web.Models.Cohort
 {
     public class ConfirmEmployerRequest : IAuthorizationContextModel
     {
