+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/CreateCohortWithDraftApprenticeshipRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SFA.DAS.Authorization.ModelBinding;
 
-namespace SFA.DAS.ProviderCommitments.Web.Requests
+namespace SFA.DAS.ProviderCommitments.Web.Models
 {
     public class CreateCohortWithDraftApprenticeshipRequest : IAuthorizationContextModel
     {
