+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/ReservationsAddDraftApprenticeshipRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SFA.DAS.Authorization.ModelBinding;
 
-namespace SFA.DAS.ProviderCommitments.Web.Requests
+namespace SFA.DAS.ProviderCommitments.Web.Models
 {
     public class ReservationsAddDraftApprenticeshipRequest : IAuthorizationContextModel
     {
